% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R, R/01-modelMap_MapLayers-class.R
\name{add_borders}
\alias{add_borders}
\alias{add_borders.MapLayers}
\title{Add borders layer (generic)}
\usage{
add_borders(x, ...)

\method{add_borders}{MapLayers}(x, ...)
}
\arguments{
\item{x}{A \code{MapLayers} object.}

\item{...}{Further arguments passed to methods.}
}
\description{
Generic for adding political/country borders to a map plot.
}
\details{
For \code{MapLayers}, \code{add_borders()} draws political borders.
}
