% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R, R/01-modelMap_MapLayers-class.R
\name{add_center_line}
\alias{add_center_line}
\alias{add_center_line.MapLayers}
\title{Add center line (generic)}
\usage{
add_center_line(x, ...)

\method{add_center_line}{MapLayers}(x, ...)
}
\arguments{
\item{x}{A \code{MapLayers} object.}

\item{...}{Further arguments passed to methods.}
}
\description{
Generic for adding a center meridian line to a map plot.
}
\details{
For \code{MapLayers}, \code{add_center_line()} draws a vertical
  center meridian at longitude 0 (plot coords).
}
