% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R, R/01-modelMap_MapLayers-class.R
\name{add_coast}
\alias{add_coast}
\alias{add_coast.MapLayers}
\title{Add coastline layer (generic)}
\usage{
add_coast(x, ...)

\method{add_coast}{MapLayers}(x, ...)
}
\arguments{
\item{x}{A \code{MapLayers} object.}

\item{...}{Further arguments passed to methods.}
}
\description{
Generic for adding a coastline layer to a map plot.
}
\details{
For \code{MapLayers}, \code{add_coast()} draws coastlines.
}
