% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R, R/01-modelMap_MapLayers-class.R
\name{add_grids}
\alias{add_grids}
\alias{add_grids.MapLayers}
\title{Add grid lines layer (generic)}
\usage{
add_grids(x, ...)

\method{add_grids}{MapLayers}(x, ...)
}
\arguments{
\item{x}{A \code{MapLayers} object.}

\item{...}{Further arguments passed to methods.}
}
\description{
Generic for adding grid lines to a map plot.
}
\details{
For \code{MapLayers}, \code{add_grids()} overlays graticule lines.
}
