% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R, R/01-modelMap_MapLayers-class.R
\name{add_land}
\alias{add_land}
\alias{add_land.MapLayers}
\title{Add land layer (generic)}
\usage{
add_land(x, ...)

\method{add_land}{MapLayers}(x, ...)
}
\arguments{
\item{x}{A \code{MapLayers} object.}

\item{...}{Further arguments passed to methods.}
}
\description{
Generic for adding a land layer to a map plot.
}
\details{
For \code{MapLayers}, \code{add_land()} fills land polygons.
}
