% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R, R/01-modelMap_MapLayers-class.R
\name{add_ocean}
\alias{add_ocean}
\alias{add_ocean.MapLayers}
\title{Add ocean layer (generic)}
\usage{
add_ocean(x, ...)

\method{add_ocean}{MapLayers}(x, ...)
}
\arguments{
\item{x}{A \code{MapLayers} object.}

\item{...}{Further arguments passed to methods.}
}
\description{
Generic for adding an ocean layer to a map plot.
}
\details{
For \code{MapLayers}, \code{add_ocean()} fills the ocean area
  respecting \code{xlim}/\code{ylim} and \code{centerMap} wrapping.
}
