% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-leafletMapHelpers.R
\name{addCirclesRelativeToZoom}
\alias{addCirclesRelativeToZoom}
\title{Add Circles to existing Map}
\usage{
addCirclesRelativeToZoom(map, isoData, mapBounds, newZoom, zoom = 5)
}
\arguments{
\item{map}{leaflet map}

\item{isoData}{isoData data}

\item{mapBounds}{bounds for the lat/lng frame}

\item{newZoom}{numeric zoom factor used for lat/lng jitter and radius calculation}

\item{zoom}{numeric default zoom factor used for lat/lng jitter and radius calculation}
}
\description{
Add Circles to existing Map
}
