% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-leafletMapHelpers.R
\name{addCustomCircles}
\alias{addCustomCircles}
\title{Add Custom Circles to existing Map}
\usage{
addCustomCircles(map, isoData, style, mapBounds)
}
\arguments{
\item{map}{leaflet map}

\item{isoData}{isoData data}

\item{style}{reactive list of values to style the circles}

\item{mapBounds}{bounds for the lat/lng frame}
}
\description{
Add Custom Circles to existing Map
}
