% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{addDataToLeafletMap}
\alias{addDataToLeafletMap}
\title{Add Data To Map}
\usage{
addDataToLeafletMap(map, isoData, leafletPointValues)
}
\arguments{
\item{map}{reactive leaflet map object}

\item{isoData}{reactive isoData data}

\item{leafletPointValues}{reactive settings for points on map}
}
\description{
Add Data To Map
}
