% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMap.R
\name{addFormattedAxis}
\alias{addFormattedAxis}
\title{Add Formatted Axis}
\usage{
addFormattedAxis(axis, min, max, nLabels = 7, decPlace = 0)
}
\arguments{
\item{axis}{(character) axis to add, either "x" or "y"}

\item{min}{(numeric) position of 1st label}

\item{max}{(numeric) position of last label}

\item{nLabels}{(numeric) number of displayed labels}

\item{decPlace}{(numeric) number of the label's decimal places}
}
\description{
Add Formatted Axis
}
