% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotExport.R
\name{addGif}
\alias{addGif}
\title{Add GIF}
\usage{
addGif(gifFile, file)
}
\arguments{
\item{gifFile}{The gif file to add a slide to}

\item{file}{the file, url, or raster object or bitmap array to be added to gifFileSource}
}
\description{
Add GIF
}
