% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-processCoordinateData.R
\name{augmentData}
\alias{augmentData}
\title{Augment Data}
\usage{
augmentData(data)
}
\arguments{
\item{data}{(data.frame) data containing columns "Latitude" and "Longitude"}
}
\value{
(data.frame) augmented data
}
\description{
Augment Data
}
