% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-augmentData.R
\name{centerData}
\alias{centerData}
\title{Center Data}
\usage{
centerData(data, center = c("0th", "180th"))
}
\arguments{
\item{data}{(data.frame) data containing columns "Latitude" and "Longitude"}

\item{center}{(character) center to shift data to, either "0th" or "180th"}
}
\value{
(data.frame) data shifted to the center
}
\description{
Center Data
}
