% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{centerEstimateServer}
\alias{centerEstimateServer}
\title{Center Estimate Server}
\usage{
centerEstimateServer(id, meanCenter, sdCenter, mapType = reactiveVal("Map"))
}
\arguments{
\item{id}{namespace id}

\item{meanCenter}{(reactive) mean for center}

\item{sdCenter}{(reactive) error for center}

\item{mapType}{(reactive) type of plot, either "Map", "Time course" or
"Time intervals by cluster"}
}
\description{
Backend for center estimate module
}
