% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{centerEstimateUI}
\alias{centerEstimateUI}
\title{Center Estimate UI}
\usage{
centerEstimateUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
UI function of center estimate module
}
