% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMapData.R
\name{centerPlotData}
\alias{centerPlotData}
\title{Center Plot Data}
\usage{
centerPlotData(data, centerMap = c("Europe", "Pacific"))
}
\arguments{
\item{data}{data frame}

\item{centerMap}{(character) center of map, one of "Europe" and "Pacific"}
}
\value{
data frame
}
\description{
Center Plot Data
}
