% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMapData.R
\name{centerXPred}
\alias{centerXPred}
\title{Center XPred}
\usage{
centerXPred(XPred, XPredPac, centerMap)
}
\arguments{
\item{XPred}{(matrix) matrix of predictions}

\item{XPredPac}{(matrix) matrix of predictions for Pacific}

\item{centerMap}{(character) center of map, one of "Europe" and "Pacific"}
}
\value{
(matrix) matrix of predictions
}
\description{
Center XPred
}
