% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotSeriesExport-class.R
\name{cleanup.PlotSeriesExport}
\alias{cleanup.PlotSeriesExport}
\title{Clean up the PlotSeriesExport object}
\usage{
\method{cleanup}{PlotSeriesExport}(obj, ...)
}
\arguments{
\item{obj}{A PlotSeriesExport object.}

\item{...}{Additional arguments (not used).}
}
\value{
The modified PlotSeriesExport object with updated status.
}
\description{
This function removes the temporary directory created for the export,
and updates the status of the object.
}
