% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-ELEMENTS.R
\name{clusterMethodUI}
\alias{clusterMethodUI}
\title{Cluster Method UI}
\usage{
clusterMethodUI(ns, timeRangeInput = FALSE)
}
\arguments{
\item{ns}{namespace}

\item{timeRangeInput}{logical, if TRUE, the time range input is shown}
}
\description{
UI for the cluster method
}
