% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataExplorer.R
\name{convertLatLongWrapper}
\alias{convertLatLongWrapper}
\title{Convert Lat Long Wrapper}
\usage{
convertLatLongWrapper(data, Longitude, Latitude, CoordType, showMessage = TRUE)
}
\arguments{
\item{data}{data.frame: data}

\item{Longitude}{character: name of longitude variable}

\item{Latitude}{character: name of latitude variable}

\item{CoordType}{character: type of longitude/latitude coordinates.
One of "decimal degrees", "degrees minutes seconds" and "degrees decimal minutes"}

\item{showMessage}{(logical) if TRUE showNotification() displays success or failure}
}
\description{
Convert Lat Long Wrapper
}
