% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-createDuplicateModal.R
\name{createDuplicateModal}
\alias{createDuplicateModal}
\title{Creates a modalDialog for detectDuplicates module}
\usage{
createDuplicateModal(vars, session)
}
\arguments{
\item{vars}{dataframe column names}

\item{session}{current session}
}
\description{
Creates a modalDialog for detectDuplicates module
}
