% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-detectDuplicates-createModal.R
\name{createModal}
\alias{createModal}
\title{Creates a modalDialog for detectDuplicates module}
\usage{
createModal(vars, session)
}
\arguments{
\item{vars}{dataframe column names}

\item{session}{current session}
}
\description{
Creates a modalDialog for detectDuplicates module
}
