% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{createPchPoints}
\alias{createPchPoints}
\title{Create PCH Points}
\usage{
createPchPoints(
  pch = 16,
  width = 50,
  height = 50,
  bg = "transparent",
  col = "black",
  ...
)
}
\arguments{
\item{pch}{plotting ‘character’, i.e., symbol to use. See graphics::points for details}

\item{width}{width in pixel}

\item{height}{height in pixel}

\item{bg}{initial background colour}

\item{col}{color code or name}

\item{...}{Further graphical parameters that are passed to graphics::points()}
}
\description{
Create PCH Points
}
