% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletSettings.R
\name{customizeLeafletMap}
\alias{customizeLeafletMap}
\title{Customize Leaflet Map}
\usage{
customizeLeafletMap(leafletMap, leafletValues)
}
\arguments{
\item{leafletMap}{leaflet map}

\item{leafletValues}{map settings, e.g. scalePosition, show/hide bounds}
}
\description{
Customize leaflet map for export
}
