% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{cutAllLongStrings}
\alias{cutAllLongStrings}
\title{Cut All Strings}
\usage{
cutAllLongStrings(df, cutAt = 50)
}
\arguments{
\item{df}{(data.frame) data.frame with character and non-character columns}

\item{cutAt}{(numeric) number of characters after which to cut the entries of an character-column}
}
\description{
Cut All Strings
}
