% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{cutStrings}
\alias{cutStrings}
\title{Cut All Strings}
\usage{
cutStrings(charVec, cutAt = 50)
}
\arguments{
\item{charVec}{(character) character vector}

\item{cutAt}{(numeric) number of characters after which to cut the entries of an character-column}
}
\description{
Cut All Strings
}
