% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataExplorer.R
\name{dataExplorer}
\alias{dataExplorer}
\title{server funtion of data explorer module}
\usage{
dataExplorer(input, output, session)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
server funtion of data explorer module
}
