% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataExplorer.R
\name{dataExplorerServer}
\alias{dataExplorerServer}
\title{server funtion of data explorer module}
\usage{
dataExplorerServer(id)
}
\arguments{
\item{id}{namespace id}
}
\value{
reactive dataframe with loaded or imported isoData
}
\description{
server funtion of data explorer module
}
