% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dataExplorer.R
\name{dataExplorerUI}
\alias{dataExplorerUI}
\title{ui function of dataExplorer module}
\usage{
dataExplorerUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of dataExplorer module
}
