% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-datTable.R
\name{datTable}
\alias{datTable}
\title{Create data table}
\usage{
datTable(dat, columns = names(dat))
}
\arguments{
\item{dat}{raw data}

\item{columns}{which columns should be shown?}
}
\description{
Create data table
}
