% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{defineDecimalPlace}
\alias{defineDecimalPlace}
\title{Place decimal points for TimeCourseplot}
\usage{
defineDecimalPlace(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
UI function of for decimal place
}
