% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeData.R
\name{deleteColumnsServer}
\alias{deleteColumnsServer}
\title{Delete Columns Server}
\usage{
deleteColumnsServer(id, preparedData)
}
\arguments{
\item{id}{id of module}

\item{preparedData}{(reactive) selected data, possibly already modified}
}
\description{
Server function of the module
}
