% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-detectDuplicates.R
\name{detectDuplicatesServer}
\alias{detectDuplicatesServer}
\title{detectDuplicates module server}
\usage{
detectDuplicatesServer(id, inputData)
}
\arguments{
\item{id}{module id}

\item{inputData}{dataframe in which duplicates are searched for}
}
\description{
detectDuplicates module server
}
