% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-detectDuplicates.R
\name{detectDuplicatesUI}
\alias{detectDuplicatesUI}
\title{detectDuplicates module ui}
\usage{
detectDuplicatesUI(id, label = "Detect Duplicates")
}
\arguments{
\item{id}{module id}

\item{label}{label for actionButton which will open a modal}
}
\description{
detectDuplicates module ui
}
