% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-interactiveMap.R
\name{draw}
\alias{draw}
\title{draw Interactive Map}
\usage{
draw(
  isoData,
  zoom = 5,
  type = "1",
  scale = FALSE,
  northArrow = FALSE,
  scalePosition = "topleft",
  northArrowPosition = "bottomright",
  center = NULL
)
}
\arguments{
\item{isoData}{isoData data}

\item{zoom}{zoom}

\item{type}{map type}

\item{scale}{show scale?}

\item{northArrow}{show north arrow?}

\item{scalePosition}{position of scale}

\item{northArrowPosition}{position of north arrow}

\item{center}{where to center map (list of lat and lng)}
}
\description{
draw Interactive Map
}
