% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletSettings.R
\name{drawFittedBounds}
\alias{drawFittedBounds}
\title{Draw Fitted Bounds}
\usage{
drawFittedBounds(map, showBounds, bounds)
}
\arguments{
\item{map}{leaflet map}

\item{showBounds}{logical show/hide fitted bounds}

\item{bounds}{list of (west, east, south, north) boundaries to be drawn}
}
\description{
Draw Fitted Bounds
}
