% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-leafletMapHelpers.R
\name{drawIcons}
\alias{drawIcons}
\title{Draw Icons on Interactive Map}
\usage{
drawIcons(
  map,
  northArrow = FALSE,
  northArrowPosition = "bottomright",
  scale = FALSE,
  scalePosition = "topleft",
  logoPosition = NA
)
}
\arguments{
\item{map}{leaflet map}

\item{northArrow}{show north arrow?}

\item{northArrowPosition}{position of north arrow}

\item{scale}{show scale?}

\item{scalePosition}{position of scale}

\item{logoPosition}{character position of logo if selected, else NA}
}
\description{
Draw Icons on Interactive Map
}
