% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-leafletMapHelpers.R
\name{drawType}
\alias{drawType}
\title{Draw Type of Interactive Map}
\usage{
drawType(map, type = "1")
}
\arguments{
\item{map}{leaflet map}

\item{type}{map type}
}
\description{
Draw Type of Interactive Map
}
