% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-estimateMap.R
\name{estimateMap}
\alias{estimateMap}
\title{Estimates spatial average model with (optional) random effects (GAMM /Generalized Additive Mixed Model)}
\usage{
estimateMap(
  data,
  independent,
  Longitude,
  Latitude,
  Site = "",
  independentUncertainty = "",
  burnin = 500,
  iter = 2000,
  nChains = 1,
  K = 50,
  Bayes = FALSE,
  CoordType = "decimal degrees",
  smoothConst = 1,
  penalty = 2,
  splineType = 2,
  outlier = FALSE,
  outlierValue = 4,
  outlierD = FALSE,
  outlierValueD = 4,
  restriction = c(-90, 90, -180, 180),
  correctionPac = FALSE,
  sdVar = FALSE,
  thinning = 2
)
}
\arguments{
\item{data}{data.frame: data}

\item{independent}{character: name of independent variable}

\item{Longitude}{character: name of longitude variable}

\item{Latitude}{character: name of latitude variable}

\item{Site}{character: name of site variable (optional)}

\item{independentUncertainty}{character: uncertainty of independent variable in sd (optional)}

\item{burnin}{integer: number of burn-in iterations for Bayesian model (default = 500)}

\item{iter}{integer: number of iterations for Bayesian model (default = 2000)}

\item{nChains}{integer: number of chains for Bayesian model (default = 1)}

\item{K}{integer: number of basis functions for tprs (thin plate regression spline)}

\item{Bayes}{boolean: Bayesian model TRUE/FALSE?}

\item{CoordType}{character: type of longitude/latitude coordinates.
One of "decimal degrees", "degrees minutes seconds" and "degrees decimal minutes"}

\item{smoothConst}{numeric: adjust smoothing parameter (> 0) for Bayesian model (optional)}

\item{penalty}{numeric: 1 for constant extrapolation, 2 for linear extrapolation}

\item{splineType}{numeric: 1 for classical tprs, 2 for spherical spline}

\item{outlier}{boolean: model outlier removal TRUE/FALSE}

\item{outlierValue}{numeric: if outlier removal is TRUE, threshold for removals in sd}

\item{outlierD}{boolean: data outlier removal TRUE/FALSE}

\item{outlierValueD}{numeric: if outlierD removal is TRUE, threshold for removals in sd}

\item{restriction}{numeric vector: spatially restricts model data 4 entries for latitude (min/max) and longitude(min/max)}

\item{correctionPac}{boolean: correction (data augmentation) for pacific centering}

\item{sdVar}{boolean: variable standard deviation}

\item{thinning}{numeric: mcmc thinning for bayesian models}
}
\description{
Estimates spatial average model with (optional) random effects (GAMM /Generalized Additive Mixed Model)
}
\examples{
\dontrun{
#load data
data <- readRDS(system.file("extData", "exampleData.Rds", package = "MpiIsoApp"))
# estimate model-map
map <- estimateMap(data = data, independent = "d13C", Longitude = "longitude",
Latitude = "latitude", Site = "site")
# Plot the map
plotMap(model = map)

# Alternative: use app
shiny::runApp(paste0(system.file(package = "MpiIsoApp"),"/app"))

}
}
