% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-estimateMap.R
\name{estimateMap3DKernel}
\alias{estimateMap3DKernel}
\title{Estimates spatio-temporal kernel density model}
\usage{
estimateMap3DKernel(
  data,
  Longitude,
  Latitude,
  DateOne,
  DateTwo,
  center = c("Europe", "Pacific"),
  independent = NULL,
  DateType = "Interval",
  CoordType = "decimal degrees",
  Weighting = NULL,
  clusterMethod = NULL,
  nClust = 5,
  nClustRange = c(2, 10),
  kMeansAlgo = "Hartigan-Wong",
  clusterTimeRange = c(0, 1000),
  modelUnc = FALSE,
  dateUnc = "point",
  restriction = c(-90, 90, -180, 180),
  nSim = 10,
  kdeType = "1"
)
}
\arguments{
\item{data}{data.frame: data}

\item{Longitude}{character: name of longitude variable}

\item{Latitude}{character: name of latitude variable}

\item{DateOne}{character: name of date variable 1 (lower interval point / mean / single point)}

\item{DateTwo}{character: name of date variable 2 (upper interval point / sd / )}

\item{center}{(character) center to shift data to, either "Europe" or "Pacific"}

\item{independent}{character: name of presence/absence variable (optional)}

\item{DateType}{character: one of "Interval", "Mean + 1 SD uncertainty" and "Single Point"}

\item{CoordType}{character: type of longitude/latitude coordinates.
One of "decimal degrees", "degrees minutes seconds" and "degrees decimal minutes"}

\item{Weighting}{character: name of weighting variable}

\item{clusterMethod}{character: cluster method}

\item{nClust}{numeric: how many clusters}

\item{nClustRange}{numeric: range of potential mclust cluster}

\item{kMeansAlgo}{character: kmeans algorithm as in stats:kmeans}

\item{clusterTimeRange}{numeric vector: time range of cluster}

\item{modelUnc}{boolean: Include dating uncertainty}

\item{dateUnc}{character: one of "uniform", "normal", "point"}

\item{restriction}{numeric vector: spatially restricts model data 4 entries for latitude (min/max) and longitude(min/max)}

\item{nSim}{numeric: number of bootstrap samples}

\item{kdeType}{character: "1" for correlated bandwidth, "2" for diagonal bandwidth, "3" for diagonal, equal long/lat bandwidth}
}
\description{
Estimates spatio-temporal kernel density model
}
\examples{
\dontrun{
# load data
data <- readRDS(system.file("extData", "exampleData.Rds", package = "MpiIsoApp"))
# estimate model-map
map <- estimateMap3D(data = data, independent = "d13C", Longitude = "longitude",
Latitude = "latitude", DateOne = "dateLower", DateTwo = "dateUpper", Site = "site")
# Plot the map
plotMap3D(model = map, time = median(data$dateLower, na.rm = TRUE))
}
}
