% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataExport.R
\name{exportCSV}
\alias{exportCSV}
\title{Export to csv}
\usage{
exportCSV(file, dat, colseparator, decseparator)
}
\arguments{
\item{file}{filename}

\item{dat}{data.frame}

\item{colseparator}{column seperator}

\item{decseparator}{decimal seperator}
}
\description{
Export to csv
}
