% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataExport.R
\name{exportFilename}
\alias{exportFilename}
\title{Filename of Export}
\usage{
exportFilename(filename = "isotopeData", fileending)
}
\arguments{
\item{filename}{name of file}

\item{fileending}{character csv or xlsx}
}
\description{
Filename of Export
}
