% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotSeriesExport-class.R
\name{exportMapR.PlotSeriesExport}
\alias{exportMapR.PlotSeriesExport}
\title{Export the series of plots for MapR}
\usage{
\method{exportMapR}{PlotSeriesExport}(obj, file, input)
}
\arguments{
\item{obj}{A PlotSeriesExport object.}

\item{file}{The name of the output ZIP file.}

\item{input}{A reactive input object containing the group, variable, and measure names.}
}
\value{
The modified PlotSeriesExport object with updated status.
}
\description{
This function exports the series of plots generated by the PlotSeriesExport object
to a ZIP file containing the images and a JSON file listing the images.
}
