% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotSeriesExport-class.R
\name{exportSeries.PlotSeriesExport}
\alias{exportSeries.PlotSeriesExport}
\title{Export the series of plots to files or a GIF}
\usage{
\method{exportSeries}{PlotSeriesExport}(
  obj,
  file,
  modelType,
  typeOfSeries,
  fpsGif = 1,
  reverseGif = FALSE,
  ...
)
}
\arguments{
\item{obj}{A PlotSeriesExport object.}

\item{file}{The name of the output file (ZIP or GIF).}

\item{modelType}{A string indicating the type of model (e.g., "local-average", "spread", etc.) for naming conventions.}

\item{typeOfSeries}{A string indicating the type of series to export ("onlyZip", "onlyGif", or "gifAndZip").}

\item{fpsGif}{The frames per second for the GIF (default is 1).}

\item{reverseGif}{A boolean indicating whether to reverse the order of the GIF frames (default is FALSE).}

\item{...}{Additional arguments (not used).}
}
\value{
The modified PlotSeriesExport object with updated status.
}
\description{
This function handles the export of the generated plots
to either a ZIP file or a GIF, depending on the specified type of series.
}
