% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelMap_UIHelpers.R
\name{extractCenterEstimates}
\alias{extractCenterEstimates}
\title{Extract Center Estimates}
\usage{
extractCenterEstimates(XPredCenter, digits = 5, batch = FALSE)
}
\arguments{
\item{XPredCenter}{(data.frame) data.frame containing estimates around a center}

\item{digits}{(numeric) significant digits}

\item{batch}{(logical) set TRUE if used for batch estimates}
}
\description{
Extract Center Estimates
}
