% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-datTable.R
\name{extractChoicesFromIsomemoApi}
\alias{extractChoicesFromIsomemoApi}
\title{Extract Choices From Isomemo Api}
\usage{
extractChoicesFromIsomemoApi(apiOutput)
}
\arguments{
\item{apiOutput}{output from the isomemo api}
}
\description{
Extract Choices From Isomemo Api
}
