% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{extractGridLength}
\alias{extractGridLength}
\title{Extract Grid Length}
\usage{
extractGridLength(latitude, longitude, digits = 5)
}
\arguments{
\item{latitude}{(numeric) latitude}

\item{longitude}{(numeric) longitude}

\item{digits}{(numeric) significant digits}
}
\description{
Extract Grid Length
}
