% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMapData.R
\name{extractMaskDraw}
\alias{extractMaskDraw}
\title{Extract Mask Draw}
\usage{
extractMaskDraw(dat, maskRadius, XPred)
}
\arguments{
\item{dat}{(data.frame) data frame}

\item{maskRadius}{(numeric) mask radius}

\item{XPred}{(matrix) matrix of predictions}
}
\value{
(numeric) mask draw
}
\description{
Extract Mask Draw
}
