% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeData.R
\name{extractMergeNotification}
\alias{extractMergeNotification}
\title{Extract Merge Notification}
\usage{
extractMergeNotification(warningsList, errorsList)
}
\arguments{
\item{warningsList}{(list) merge warnings}

\item{errorsList}{(list) merge errors}
}
\description{
Extract Merge Notification
}
