% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeData.R
\name{extractTableIds}
\alias{extractTableIds}
\title{Extract Table IDs}
\usage{
extractTableIds(namesOfTables)
}
\arguments{
\item{namesOfTables}{(character) names of loaded tables, often url's to the table file, something
like "https://pandoradata.earth/dataset/.../download/afriarch-isotopic-dataset.xlsx"}
}
\value{
(character) short internal table names
}
\description{
Extract Table IDs
}
