% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{extractZoomFromLongRange}
\alias{extractZoomFromLongRange}
\title{Extract Zoom From Long Range}
\usage{
extractZoomFromLongRange(rangeLongitude, mapCentering)
}
\arguments{
\item{rangeLongitude}{(numeric) range of longitude vector}

\item{mapCentering}{(character) centering of the map, either "Europe" or "Pacific"}
}
\description{
Extract Zoom From Long Range
}
