% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMap.R
\name{filled.contour2}
\alias{filled.contour2}
\title{Filled Countour 2
Wrapper for contour plot}
\usage{
filled.contour2(
  x = seq(0, 1, length.out = nrow(z)),
  y = seq(0, 1, length.out = ncol(z)),
  z,
  contourType = c("filled.contour", "contour"),
  xlim = range(x, finite = TRUE),
  ylim = range(y, finite = TRUE),
  zlim = range(z, finite = TRUE),
  levels = pretty(zlim, nlevels),
  nlevels = 20,
  showScale = TRUE,
  color.palette = cm.colors,
  col = color.palette(length(levels) - 1),
  plot.title,
  plot.axes,
  key.title,
  key.axes,
  asp = NA,
  xaxs = "i",
  yaxs = "i",
  las = 1,
  axes = TRUE,
  frame.plot = axes,
  ...
)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{z}{z values}

\item{contourType}{one of "filled.contour" or "contour"}

\item{xlim}{x limits}

\item{ylim}{y limits}

\item{zlim}{z limits}

\item{levels}{levels}

\item{nlevels}{number of levels}

\item{showScale}{show colour scale}

\item{color.palette}{color palette}

\item{col}{colors}

\item{plot.title}{plot title}

\item{plot.axes}{plot axes}

\item{key.title}{key title}

\item{key.axes}{key axes}

\item{asp}{aspect ratio}

\item{xaxs}{x axis style}

\item{yaxs}{y axis style}

\item{las}{label style}

\item{axes}{show axes}

\item{frame.plot}{frame plot}

\item{...}{additional arguments}
}
\description{
Filled Countour 2
Wrapper for contour plot
}
