% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-savedMapsModule.R
\name{filterCoordCircle}
\alias{filterCoordCircle}
\title{Filter Coord Circle}
\usage{
filterCoordCircle(fullGrid, lat, long, radius)
}
\arguments{
\item{fullGrid}{grid of coordinates}

\item{lat}{(numeric) latitude}

\item{long}{(numeric) longitude}

\item{radius}{(numeric) radius}
}
\description{
Filter full grid of coordinates to circle area
}
