% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-savedMapsModule.R
\name{filterCoordRectangle}
\alias{filterCoordRectangle}
\title{Filter Coord Rectangle}
\usage{
filterCoordRectangle(fullGrid, lat, long, latLength, longLength)
}
\arguments{
\item{fullGrid}{grid of coordinates}

\item{lat}{(numeric) latitude of center}

\item{long}{(numeric) longitude of center}

\item{latLength}{(numeric) length between latitudes}

\item{longLength}{(numeric) length between longitudes}
}
\description{
Filter full grid of coordinates to square area
}
