% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-savedMapsModule.R
\name{filterCoordSquare}
\alias{filterCoordSquare}
\title{Filter Coord Square}
\usage{
filterCoordSquare(fullGrid, lat, long, length)
}
\arguments{
\item{fullGrid}{grid of coordinates}

\item{lat}{(numeric) latitude of center}

\item{long}{(numeric) longitude of center}

\item{length}{(numeric) side length of square}
}
\description{
Filter full grid of coordinates to square area
}
