% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMapData.R
\name{filterT}
\alias{filterT}
\title{Filter Time}
\usage{
filterT(data, addU, time)
}
\arguments{
\item{data}{(data.frame) data to be filtered by time containing columns "Date" and "Uncertainty"}

\item{addU}{(numeric) additional uncertainty to be added to time frame}

\item{time}{(numeric) time for filter}
}
\value{
(data.frame) filtered data
}
\description{
Filter Time
}
