% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-findDuplicates.R
\name{findDuplicates}
\alias{findDuplicates}
\title{Finds duplicates in data using user specified similarity rules}
\usage{
findDuplicates(data, userSimilaritySelection)
}
\arguments{
\item{data}{dataframe in which duplicates are searched for}

\item{userSimilaritySelection}{dataframe containing similarity rules for each column to be considered}
}
\description{
Finds duplicates in data using user specified similarity rules
}
