% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-logging.R
\name{format_bytes}
\alias{format_bytes}
\title{Convert bytes to human-readable units}
\usage{
format_bytes(bytes)
}
\arguments{
\item{bytes}{(numeric) bytes}
}
\description{
Convert bytes to human-readable units
}
