% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{formatColumnNames}
\alias{formatColumnNames}
\title{Format Column Names}
\usage{
formatColumnNames(vNames, isTest = FALSE)
}
\arguments{
\item{vNames}{(character) names of the imported data's columns}

\item{isTest}{(logical) set TRUE if function is used in tests}
}
\description{
Replaces all not alpha-numeric characters in the names of columns with a dot.
}
