% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{formatTimeCourseUI}
\alias{formatTimeCourseUI}
\title{Format Time Course UI}
\usage{
formatTimeCourseUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
UI function for formatting of the time course plot
}
