% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotSeriesExport-class.R
\name{generate.PlotSeriesExport}
\alias{generate.PlotSeriesExport}
\title{Generate the series of plots and save them to files
This function iterates over the specified time points,
calls the plotting function for each time point,
and saves the plots in the specified format.}
\usage{
\method{generate}{PlotSeriesExport}(obj, ...)
}
\arguments{
\item{obj}{A PlotSeriesExport object.}

\item{...}{Additional arguments (not used).}
}
\value{
The modified PlotSeriesExport object with updated status.
}
\description{
Generate the series of plots and save them to files
This function iterates over the specified time points,
calls the plotting function for each time point,
and saves the plots in the specified format.
}
