% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotExport.R
\name{generateGif}
\alias{generateGif}
\title{Generate GIF}
\usage{
generateGif(gifFile = "animated.gif", files, fps = 1)
}
\arguments{
\item{gifFile}{The gif file to create}

\item{files}{a list of files, url's, or raster objects or bitmap arrays}

\item{fps}{frames per second}
}
\description{
Generate GIF
}
