% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dataSearchMode.R
\name{getAppTitle}
\alias{getAppTitle}
\title{Get App Title}
\usage{
getAppTitle()
}
\description{
If the full app is accessible (if Sys.getenv("ISOMEMO_HIDE_MODELLING") == ""),
sets the title of the app to "Data search & Spatiotemporal modeling"; and else to
"Data search", the modeling feature is hidden in that case.
}
