% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-mapOperations.R
\name{getBinaryPredictions}
\alias{getBinaryPredictions}
\title{Get Binary Predictions}
\usage{
getBinaryPredictions(XPred1, XPred2)
}
\arguments{
\item{XPred1}{prediction matrix of first map}

\item{XPred2}{prediction matrix of second map}
}
\value{
prediction matrix of resulting map
}
\description{
Get Binary Predictions
}
