% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictNearestCluster.R
\name{getCentroid}
\alias{getCentroid}
\title{Get centroid of clusters}
\usage{
getCentroid(dataC, cluster, centers, removeClusterCol = FALSE)
}
\arguments{
\item{dataC}{Filtered data}

\item{removeClusterCol}{Remove cluster column from data}

\item{clust_centroid}{Centroid of clusters}
}
\value{
Filtered data with centroid
}
\description{
Get centroid of clusters
}
