% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{getColourPal}
\alias{getColourPal}
\title{Get Colour Palette}
\usage{
getColourPal(values)
}
\arguments{
\item{values}{possible values that can be mapped, e.g. isoData$source}
}
\description{
Get colour palette for the points and the legend
}
