% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-savedMapsModule.R
\name{getCoordCenter}
\alias{getCoordCenter}
\title{Get Coord Center}
\usage{
getCoordCenter(upperLeftLat, upperLeftLong, lowerRightLat, lowerRightLong)
}
\arguments{
\item{upperLeftLat}{(numeric) latitude of upper left point}

\item{upperLeftLong}{(numeric) longitude of upper left point}

\item{lowerRightLat}{(numeric) latitude of lower right point}

\item{lowerRightLong}{(numeric) longitude of lower right point}
}
\description{
Get center of rectangle defined by upper left and lower right coordinates
}
