% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMap.R
\name{getCostSpeed}
\alias{getCostSpeed}
\title{Get Speed or Cost object for Cost-Surface and least-cost path}
\usage{
getCostSpeed(XPred, MinMax)
}
\arguments{
\item{XPred}{(data.frame) data frame containing the spread function estimates}

\item{MinMax}{(string) are we estimating min or max}
}
\value{
speed object
}
\description{
Get Speed or Cost object for Cost-Surface and least-cost path
}
