% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-locationFields.R
\name{getDefaultCoordColumn}
\alias{getDefaultCoordColumn}
\title{Get Default Coord Column}
\usage{
getDefaultCoordColumn(columnNames, tryPattern = c("latitude", "^lat$"))
}
\arguments{
\item{columnNames}{(character) column names of loaded data}

\item{tryPattern}{(character) pattern that should be matched with column names ordered after
priority}
}
\description{
Get Default Coord Column
}
