% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{getDefaultZError}
\alias{getDefaultZError}
\title{Get Default Z Error}
\usage{
getDefaultZError(estType, range)
}
\arguments{
\item{estType}{(character) type of estimate}

\item{range}{(numeric) range from model output}
}
\description{
Get Default Z Error
}
