% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotExport.R
\name{getFileName}
\alias{getFileName}
\title{Get File Name}
\usage{
getFileName(plotType, isTimeSeries, i = NULL)
}
\arguments{
\item{plotType}{(character) plot specification}

\item{isTimeSeries}{(logical) if TRUE, set file names for a series of plots}

\item{i}{(numeric) number of i-th plot of a series of plots}
}
\description{
Get File Name
}
