% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-savedMapsModule.R
\name{getFullCoordGrid}
\alias{getFullCoordGrid}
\title{Get Full Coord Grid}
\usage{
getFullCoordGrid(gridLength)
}
\arguments{
\item{gridLength}{length between two points}
}
\description{
Get Full Coord Grid
}
