% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{getHTMLFromPath}
\alias{getHTMLFromPath}
\title{Get HTML from path}
\usage{
getHTMLFromPath(paths)
}
\arguments{
\item{paths}{named list of paths to the files containing an icon, names of the items are used as
labels}
}
\description{
Creates the HTML string that defines the legend
}
