% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{getNrow}
\alias{getNrow}
\title{get nRow}
\usage{
getNrow(headOnly, type, n = 4)
}
\arguments{
\item{headOnly}{(logical) if TRUE, set maximal number of rows to n}

\item{type}{(character) file type}

\item{n}{(numeric) maximal number of rows if headOnly}
}
\description{
get nRow
}
