% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMap.R
\name{getPColor}
\alias{getPColor}
\title{Get Point Color}
\usage{
getPColor(data, cluster, palName, pColor)
}
\arguments{
\item{data}{(data.frame) if cluster, different ids in the column "cluster" will receive a different colour}

\item{cluster}{(logical) if TRUE, return color for cluster}

\item{palName}{(character) name of the color palette}

\item{pColor}{single color or vector of colors that is returned if \code{cluster == FALSE}}
}
\value{
single color or vector of colors
}
\description{
Get Point Color
}
