% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{getPointSize}
\alias{getPointSize}
\title{Get Point Size}
\usage{
getPointSize(df, columnForPointSize, sizeFactor = 1)
}
\arguments{
\item{df}{(data.frame) loaded data}

\item{columnForPointSize}{(character) name of the column that determines the point size}

\item{sizeFactor}{(numeric) general factor for point size}
}
\description{
Get point size in pixel
}
