% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{getPointSymbols}
\alias{getPointSymbols}
\title{Get Point Size}
\usage{
getPointSymbols(df, columnForPointSymbols, symbols = unlist(pchChoices()))
}
\arguments{
\item{df}{(data.frame) loaded data}

\item{columnForPointSymbols}{(character) name of the column that determines the point symbol}

\item{symbols}{(numeric) selected symbols}
}
\description{
Get Point Size
}
