% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{getSheetSelection}
\alias{getSheetSelection}
\title{Get Sheet Selection}
\usage{
getSheetSelection(filepath)
}
\arguments{
\item{filepath}{(character) url or path}
}
\description{
Get Sheet Selection
}
