% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelMap_UIHelpers.R
\name{getZvalues}
\alias{getZvalues}
\title{Get Z Values}
\usage{
getZvalues(
  estimationType,
  model,
  mapType,
  factor = 3,
  IndSelect = NULL,
  buffer = 0.01
)
}
\arguments{
\item{estimationType}{(character) type of estimate}

\item{model}{(list) model output}

\item{mapType}{(character) type of map, either "Map" or "Time course"}

\item{factor}{(numeric) factor applied to estimates}

\item{IndSelect}{(character) select category in case of categorical model}

\item{buffer}{(numeric) buffer applied to the mean range, e.g. 0.01}
}
\description{
Get Z Values
}
