% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{getZValuesMapSim}
\alias{getZValuesMapSim}
\title{Get Z Values Map Sim}
\usage{
getZValuesMapSim(estimationType, model, mapType, factor = 1)
}
\arguments{
\item{estimationType}{(character) type of estimate}

\item{model}{(list) model output}

\item{mapType}{(character) type of map, either "Map" or "Time course"}

\item{factor}{(numeric) factor applied to estimates}
}
\description{
Get Z Values Map Sim
}
