% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{importDataUI}
\alias{importDataUI}
\title{Data import module}
\usage{
importDataUI(id, label = "Import Data")
}
\arguments{
\item{id}{id of module}

\item{label}{label of button}
}
\description{
Displays a button which opens a import dialog when clicked
}
