% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{importDataServer}
\alias{importDataServer}
\title{Server function for data import}
\usage{
importDataServer(
  id,
  rowNames = NULL,
  colNames = NULL,
  customWarningChecks = list(),
  customErrorChecks = list()
)
}
\arguments{
\item{id}{namespace id}

\item{rowNames}{(reactive) use this for rownames of imported data}

\item{colNames}{(reactive) use this for colnames of imported data}

\item{customWarningChecks}{list of reactive functions which will be executed after importing of data.
functions need to return TRUE if check is successful or a character with a warning otherwise.}

\item{customErrorChecks}{list of reactive functions which will be executed after importing of data.
functions need to return TRUE if check is successful or a character with a warning otherwise.}
}
\description{
Backend for data import module
}
