% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-interactiveMap.R
\name{interactiveMap}
\alias{interactiveMap}
\title{server funtion of interactive map module}
\usage{
interactiveMap(input, output, session, isoData)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{isoData}{data}
}
\description{
server funtion of interactive map module
}
