% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-interactiveMap.R
\name{interactiveMapUI}
\alias{interactiveMapUI}
\title{ui function of interactiveMap module}
\usage{
interactiveMapUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of interactiveMap module
}
