% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dataSearchMode.R
\name{isOnlyDataSearchMode}
\alias{isOnlyDataSearchMode}
\title{Is Only Data SEARCH}
\usage{
isOnlyDataSearchMode()
}
\value{
(logical) TRUE if "only-data-search mode" is active, else FALSE for full version of the app.
}
\description{
Checks if "only-data-search mode" is active.
}
