% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-jitter-latlong.R
\name{jitter_latlong}
\alias{jitter_latlong}
\title{Jitter latitude and longitude coordinates}
\usage{
jitter_latlong(coord, type = c("lat", "long"), latitude, km = 1)
}
\arguments{
\item{coord}{The coordinate in degrees}

\item{type}{Whether the coordinate is latitude or longitude}

\item{latitude}{If the coordinate is longitude, then the latitude to use for
calculating the maximum amount to jitter.}

\item{km}{The maximum number of kilometers to jitter a point plus or minus.}
}
\value{
The jittered coordinate in degrees.
}
\description{
Given a latitude or longitude coordinate, return that coordinate jittered
plus or minus a certain maximum amount. This function assumes a spherical
earth when calculating the maximum amount to jitter.
}
\examples{
jitter_latlong(-73, type = "long", lat = 43, km = 1)
jitter_latlong(42, type = "lat", km = 1)
}
\seealso{
\code{\link{length_of_degree}}
}
