% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictNearestCluster.R
\name{joinCentroid}
\alias{joinCentroid}
\title{Join centroid to data}
\usage{
joinCentroid(data, dataC)
}
\arguments{
\item{data}{Data}

\item{dataC}{Filtered data with centroid}
}
\description{
Join centroid to data
}
