% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictNearestCluster.R
\name{joinCentroidData}
\alias{joinCentroidData}
\title{Join centroid data to data}
\usage{
joinCentroidData(data, dataC)
}
\arguments{
\item{data}{Data}

\item{dataC}{Filtered data with centroid}
}
\value{
Data with centroid
}
\description{
Join centroid data to data
}
