% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeData.R
\name{joinColumnsServer}
\alias{joinColumnsServer}
\title{Join Columns Server}
\usage{
joinColumnsServer(id, preparedData)
}
\arguments{
\item{id}{id of module}

\item{preparedData}{(reactive) (reactive) selected data, possibly already modified}
}
\description{
Server function of the module
}
