% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-leafletMapHelpers.R
\name{leafletAddCircles}
\alias{leafletAddCircles}
\title{Add Circles to existing Map}
\usage{
leafletAddCircles(map, isoData, style, newZoom, mapBounds)
}
\arguments{
\item{map}{leaflet map}

\item{isoData}{isoData data}

\item{style}{reactive list of values to style the circles}

\item{newZoom}{numeric zoom factor used for lat/lng jitter and radius calculation}

\item{mapBounds}{bounds for the lat/lng frame}
}
\description{
Add Circles to existing Map
}
