% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletDataStyle.R
\name{leafletDataStyle}
\alias{leafletDataStyle}
\title{server funtion of leaflet settings module}
\usage{
leafletDataStyle(input, output, session, isoData)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{isoData}{reactive isoData table}
}
\description{
server funtion of leaflet settings module
}
