% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletDataStyle.R
\name{leafletDataStyleUI}
\alias{leafletDataStyleUI}
\title{ui function of leaflet settings module}
\usage{
leafletDataStyleUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of leaflet settings module
}
