% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletExport.R
\name{leafletExport}
\alias{leafletExport}
\title{server funtion of leaflet map export module}
\usage{
leafletExport(
  input,
  output,
  session,
  isoData,
  zoom,
  center,
  width,
  height,
  leafletValues
)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
server funtion of leaflet map export module
}
