% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletExport.R
\name{leafletExportButton}
\alias{leafletExportButton}
\title{ui function of leaflet map export module}
\usage{
leafletExportButton(id)
}
\arguments{
\item{id}{namespace}
}
\description{
ui function of leaflet map export module
}
