% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletSettings.R
\name{leafletSettings}
\alias{leafletSettings}
\title{server funtion of leaflet settings module}
\usage{
leafletSettings(input, output, session, zoom, center)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{zoom}{map zoom}

\item{center}{where to center map (list of lat and lng)}
}
\description{
server funtion of leaflet settings module
}
