% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletSettings.R
\name{leafletSettingsUI}
\alias{leafletSettingsUI}
\title{ui function of leaflet settings module}
\usage{
leafletSettingsUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of leaflet settings module
}
