% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-jitter-latlong.R
\name{length_of_degree}
\alias{length_of_degree}
\title{Length of a degree of latitude or longitude}
\usage{
length_of_degree(degree, type = c("lat", "long"))
}
\arguments{
\item{degree}{The degree to calculate the length for}

\item{type}{Whether to return the length of a degree for latitude or longitude}
}
\value{
Length of the degree in kilometers
}
\description{
Calculates the length of a degree of latitude or longitude in kilometers,
assuming an spherical earth.
}
