% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{loadDataWrapper}
\alias{loadDataWrapper}
\title{Load Data Wrapper}
\usage{
loadDataWrapper(
  values,
  filepath,
  filename,
  colNames,
  type,
  sep,
  dec,
  withRownames,
  headOnly,
  customWarningChecks,
  customErrorChecks
)
}
\arguments{
\item{filepath}{url or path}

\item{filename}{url or file name}

\item{colNames}{(reactive) use this for colnames of imported data}

\item{type}{(character) file type input}

\item{sep}{(character) column separator input}

\item{dec}{(character) decimal separator input}

\item{withRownames}{(logical) contains rownames input}

\item{headOnly}{(logical) load only head (first n rows) of file}

\item{customWarningChecks}{list of reactive functions which will be executed after importing of data.
functions need to return TRUE if check is successful or a character with a warning otherwise.}

\item{customErrorChecks}{list of reactive functions which will be executed after importing of data.
functions need to return TRUE if check is successful or a character with a warning otherwise.}

\item{value}{(list) list with import specifications}
}
\description{
Load Data Wrapper
}
