% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-dataHelpers.R
\name{locationFieldsServer}
\alias{locationFieldsServer}
\title{Location Fields Server}
\usage{
locationFieldsServer(id, dataRaw)
}
\arguments{
\item{id}{namespace id}

\item{dataRaw}{(reactive) raw data import}
}
\description{
Backend for the module
}
