% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-locationFields.R
\name{locationFieldsUI}
\alias{locationFieldsUI}
\title{Location Fields UI}
\usage{
locationFieldsUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
UI function for the module
}
