% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{mapSectionUI}
\alias{mapSectionUI}
\title{Map Section UI}
\usage{
mapSectionUI(id, label)
}
\arguments{
\item{id}{id of module}

\item{label}{label}
}
\description{
UI of the Map Section module
}
