% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeImports.R
\name{mergeImportsServer}
\alias{mergeImportsServer}
\title{Merge Data Server}
\usage{
mergeImportsServer(id, mergeList)
}
\arguments{
\item{id}{id of module}

\item{mergeList}{(list) list of data to be merged}
}
\description{
Server function of the merge data module
}
