% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeViaCommand.R
\name{mergeViaCommandServer}
\alias{mergeViaCommandServer}
\title{Merge Command Server}
\usage{
mergeViaCommandServer(id, mergeCommandAuto)
}
\arguments{
\item{id}{id of module}

\item{mergeCommandAuto}{(character) automatically generated command for data merge}
}
\description{
Server function of the merge Command module
}
