% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeViaCommand.R
\name{mergeViaCommandUI}
\alias{mergeViaCommandUI}
\title{Merge Via Command UI}
\usage{
mergeViaCommandUI(id)
}
\arguments{
\item{id}{id of module}
}
\description{
UI of the merge Via command module
}
