% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeViaUI.R
\name{mergeViaUIServer}
\alias{mergeViaUIServer}
\title{Merge Via UI Server}
\usage{
mergeViaUIServer(id, tableXData, tableYData, tableXId, tableYId)
}
\arguments{
\item{id}{id of module}

\item{tableXData}{(data.frame) data to be merged}

\item{tableYData}{(data.frame) data to be merged}

\item{tableXId}{(character) internal table name, see \link{extractTableIds}}

\item{tableYId}{(character) internal table name, see \link{extractTableIds}}
}
\description{
Server function of the merge via UI module
}
