% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelLinkUI.R
\name{modelLinkUI}
\alias{modelLinkUI}
\title{ui function to display a download link}
\usage{
modelLinkUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function to display a download link
}
