% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelResults2D.R
\name{modelResults2D}
\alias{modelResults2D}
\title{server function of model Results module}
\usage{
modelResults2D(input, output, session, isoData, savedMaps, fruitsData, config)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{isoData}{data}

\item{savedMaps}{saved Maps}

\item{fruitsData}{data for export to FRUITS}

\item{config}{(list) list of configuration parameters}
}
\description{
server function of model Results module
}
