% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelResults2DKernel.R
\name{modelResults2DKernelUI}
\alias{modelResults2DKernelUI}
\title{ui function of modelResults2DKernel module}
\usage{
modelResults2DKernelUI(id, title = "", asFruitsTab = FALSE)
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}

\item{asFruitsTab}{show as part of the FRUITS app}
}
\description{
ui function of modelResults2DKernel module
}
