% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelResults2D.R
\name{modelResults2DUI}
\alias{modelResults2DUI}
\title{ui function of modelResults2D module}
\usage{
modelResults2DUI(id, title = "", asFruitsTab = FALSE)
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}

\item{asFruitsTab}{show as part of the FRUITS app}
}
\description{
ui function of modelResults2D module
}
