% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelResults3DKernel.R
\name{modelResults3DKernelUI}
\alias{modelResults3DKernelUI}
\title{ui function of modelResults3DKernel module}
\usage{
modelResults3DKernelUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of modelResults3DKernel module
}
