% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelResults3D.R
\name{modelResults3DUI}
\alias{modelResults3DUI}
\title{ui function of modelResults3D module}
\usage{
modelResults3DUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of modelResults3D module
}
