% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelResultsAssign.R
\name{modelResultsAssign}
\alias{modelResultsAssign}
\title{server function of model Results module}
\usage{
modelResultsAssign(input, output, session, isoData)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{isoData}{data}
}
\description{
server function of model Results module
}
