% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelResultsAssign.R
\name{modelResultsAssignUI}
\alias{modelResultsAssignUI}
\title{ui function of modelResultsAssignR module}
\usage{
modelResultsAssignUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of modelResultsAssignR module
}
