% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-mapDiff.R
\name{modelResultsDiffUI}
\alias{modelResultsDiffUI}
\title{ui function of modelResultsDiffSim module (OperateR)}
\usage{
modelResultsDiffUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of modelResultsDiffSim module (OperateR)
}
