% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-mapSim.R
\name{modelResultsSimUI}
\alias{modelResultsSimUI}
\title{ui function of modelResultsDiffSim module}
\usage{
modelResultsSimUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of modelResultsDiffSim module
}
