% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelResultsSpread.R
\name{modelResultsSpreadUI}
\alias{modelResultsSpreadUI}
\title{ui function of modelResultsSpread module}
\usage{
modelResultsSpreadUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of modelResultsSpread module
}
