% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotExport.R
\name{nameFile}
\alias{nameFile}
\title{Name File}
\usage{
nameFile(plotType, exportType, isTimeSeries, typeOfSeries, i = NULL)
}
\arguments{
\item{plotType}{(character) plot specification}

\item{exportType}{(character) file type of exported plot}

\item{isTimeSeries}{(logical) if TRUE, set file names for a series of plots}

\item{i}{(numeric) number of i-th plot of a series of plots}
}
\description{
Name File
}
