% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-interactiveMap_NorthArrow-class.R
\name{new_NorthArrow}
\alias{new_NorthArrow}
\title{North Arrow for Interactive Maps}
\usage{
new_NorthArrow(
  position = "none",
  width = 80,
  height = 80,
  lat = NULL,
  lng = NULL,
  layerId = "northArrowIcon"
)
}
\arguments{
\item{position}{Position of the North Arrow. Options are "topright", "topleft", "bottomright", "bottomleft", "custom", or "none".}

\item{width}{Width of the North Arrow in pixels.}

\item{height}{Height of the North Arrow in pixels.}

\item{lat}{Latitude for custom positioning (required if position is "custom").}

\item{lng}{Longitude for custom positioning (required if position is "custom").}

\item{layerId}{Unique identifier for the North Arrow layer.}
}
\value{
A NorthArrow object.
}
\description{
Creates a North Arrow object for use in interactive maps.
}
