% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotSeriesExport-class.R
\name{new_PlotSeriesExport}
\alias{new_PlotSeriesExport}
\title{PlotSeriesExport Class}
\usage{
new_PlotSeriesExport(
  plotFun,
  Model,
  times,
  exportType = "png",
  modelType = "",
  width = 1280,
  height = 800
)
}
\arguments{
\item{plotFun}{A function that generates a plot given a model and a time point.}

\item{Model}{The model object that the plotting function will use.}

\item{times}{A vector of time points at which to generate the plots.}

\item{exportType}{The type of file to export the plots as (e.g., "png", "jpeg", "pdf", etc.).}

\item{modelType}{A string indicating the type of model (e.g., "local-average", "spread", etc.) for naming conventions.}

\item{width}{The width of the exported plots in pixels.}

\item{height}{The height of the exported plots in pixels.}
}
\description{
This class is used to manage the export of a series of plots generated by a plotting function.
It allows for the generation of plots at specified time points, saving them in various formats,
and exporting them as a series of files or a GIF.
}
