% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-interactiveMap_ScaleBar-class.R
\name{new_ScaleBar}
\alias{new_ScaleBar}
\title{ScaleBar Class}
\usage{
new_ScaleBar(
  position = "none",
  zoom = NULL,
  size = 100,
  lat = NULL,
  lng = NULL
)
}
\arguments{
\item{position}{The position of the scale bar on the map. Options are "topright", "topleft", "bottomright", "bottomleft", "custom", or "none".}

\item{zoom}{The zoom level at which the scale bar is displayed. If `NULL`, it defaults to the current zoom level of the map.}

\item{size}{The size of the scale bar in pixels. Default is 100.}

\item{lat}{The latitude for a custom position of the scale bar. Required if `position` is "custom".}

\item{lng}{The longitude for a custom position of the scale bar. Required if `position` is "custom".}
}
\value{
An object of class `ScaleBar`.
}
\description{
This class represents a scale bar for interactive maps.
}
