% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-ELEMENTS.R
\name{numericInputLatAndLongServer}
\alias{numericInputLatAndLongServer}
\title{Numeric Input Lat And Long Server}
\usage{
numericInputLatAndLongServer(
  id,
  valueLat = reactive(50),
  valueLong = reactive(10)
)
}
\arguments{
\item{id}{id of module}

\item{valueLat}{(numeric) latitude}

\item{valueLong}{(numeric) longitude}
}
\description{
Server function of the numeric input of latitude and longitude module
}
