% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-ELEMENTS.R
\name{numericInputLatAndLongUI}
\alias{numericInputLatAndLongUI}
\title{Numeric Input Lat And Long UI}
\usage{
numericInputLatAndLongUI(id, label, valueLat = 50, valueLong = 10)
}
\arguments{
\item{id}{id of module}

\item{label}{label}

\item{valueLat}{(numeric) latitude}

\item{valueLong}{(numeric) longitude}
}
\description{
UI of the numeric input of latitude and longitude
}
