% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMap.R
\name{plotDS}
\alias{plotDS}
\title{Plots difference or similarity map}
\usage{
plotDS(
  XPred,
  estType = "mean",
  estQuantile = 0.95,
  type = "similarity",
  independent = "",
  rangex = range(XPred$Longitude, na.rm = TRUE),
  rangey = range(XPred$Latitude, na.rm = TRUE),
  rangez = range(XPred$Est, na.rm = TRUE),
  contourType = c("filled.contour", "contour"),
  showModel = TRUE,
  showScale = TRUE,
  ncol = 10,
  colors = "RdYlGn",
  centerMap = "Europe",
  reverseColors = FALSE,
  terrestrial = 1,
  grid = TRUE,
  showBorders = TRUE,
  arrow = TRUE,
  scale = TRUE,
  simValues = NULL,
  showValues = FALSE,
  dataCenter = NULL,
  RadiusBatch = NULL,
  titleMain = TRUE,
  titleScale = TRUE,
  setAxisLabels = FALSE,
  mainLabel = "",
  yLabel = "Latitude",
  xLabel = "Longitude",
  scLabel = "",
  northSize = 0.2,
  scalSize = 0.1,
  scaleX = 0,
  scaleY = 0.1,
  NorthX = 0.025,
  NorthY = 0.925,
  AxisSize = 1,
  AxisLSize = 1,
  pointDat = NULL
)
}
\arguments{
\item{XPred}{return object of similarityMap or createDifferenceMap functions}

\item{estType}{one of "Mean", "SE" and "Quantile". defaults to "Mean"}

\item{estQuantile}{quantile (only applicable if estType = "Quantile")}

\item{type}{one of "similarity" or "difference" determining the type of map}

\item{independent}{name of independent variable shown in plot}

\item{rangex}{range of longitude values (x axis limits)}

\item{rangey}{range of latitude values (y axis limits)}

\item{rangez}{range of estimated values (z axis limits)}

\item{contourType}{one of "filled.contour" or "contour"}

\item{showModel}{show model}

\item{showScale}{show colour scale}

\item{ncol}{number of colors for estimates}

\item{colors}{color scheme of estimates from RColorBrewer. defaults to "RdYlGn"}

\item{centerMap}{center of map, one of "Europe" and "Pacific"}

\item{reverseColors}{inverse colour scheme}

\item{terrestrial}{show only estimates on land masses (1), oceans (-1) or all (0)}

\item{grid}{show coordinate grid TRUE/FALSE}

\item{arrow}{display north arrow TRUE/FALSE}

\item{scale}{display scale TRUE/FALSE}

\item{simValues}{if showValues: list of simulated values}

\item{showValues}{boolean show values in plot?}

\item{dataCenter}{data for batch estimation}

\item{RadiusBatch}{radius}

\item{titleMain}{show main title}

\item{titleScale}{show scale title}

\item{setAxisLabels}{show axis/main/scale labels}

\item{mainLabel}{main label}

\item{yLabel}{y lab label}

\item{xLabel}{y lab label}

\item{scLabel}{scale lab labels}

\item{northSize}{size of north arrow}

\item{scalSize}{size of scale}

\item{scaleX}{scale x orientation}

\item{scaleY}{scale y orientation}

\item{NorthX}{north arrow x orientation}

\item{NorthY}{north arrow y orientation}

\item{AxisSize}{axis title font size}

\item{AxisLSize}{axis label font size}

\item{pointDat}{data frame of points to add to plot}
}
\description{
Plots difference or similarity map
}
