% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMap.R
\name{plotMap}
\alias{plotMap}
\title{Plots map of a spatial average or spread model from estimateMap() or estimateMapSpread() functions}
\usage{
plotMap(
  model,
  IndSelect = NULL,
  estType = "Mean",
  estQuantile = 0.9,
  points = TRUE,
  pointSize = 1,
  StdErr = Inf,
  rangex = range(model$data$Longitude, na.rm = TRUE),
  rangey = range(model$data$Latitude, na.rm = TRUE),
  rangez = NULL,
  limitz = NULL,
  resolution = 100,
  interior = TRUE,
  ncol = 10,
  colors = "RdYlGn",
  reverseColors = FALSE,
  colorsP = NULL,
  centerMap = "Europe",
  pointShape = 4,
  showScale = TRUE,
  textLabels = NULL,
  pointLabels = NULL,
  pointColLabels = NULL,
  fontSize = 11,
  showModel = TRUE,
  fontType = "sans",
  fontCol = "black",
  mask = FALSE,
  maskRadius = 100,
  centerX = NA,
  centerY = NA,
  Radius = NA,
  pColor = "#000000",
  dataCenter = NULL,
  RadiusBatch = NULL,
  terrestrial = 1,
  grid = TRUE,
  arrow = TRUE,
  scale = TRUE,
  mapType = "Map",
  titleMain = TRUE,
  titleScale = TRUE,
  setAxisLabels = FALSE,
  mainLabel = "",
  yLabel = "Latitude",
  xLabel = "Longitude",
  northSize = 0.2,
  scalSize = 0.1,
  scaleX = 0,
  scaleY = 0.1,
  NorthX = 0.025,
  NorthY = 0.925,
  scLabel = "",
  AxisSize = 1,
  AxisLSize = 1,
  cluster = FALSE,
  clusterCol = "Set1",
  pointDat = NULL,
  MinMax = "Min",
  nMin = 3,
  minDist = 250,
  showMinOnMap = FALSE
)
}
\arguments{
\item{model}{return object of spatial or spread model
from estimateMap() or estimateMapSpread() functions}

\item{IndSelect}{for categorical model: selected category; shifts between categories in the center}

\item{estType}{one of "Mean", "SE" and "Quantile". defaults to "Mean"}

\item{estQuantile}{quantile (only applicable if estType = "Quantile")}

\item{points}{show points on map TRUE/FALSE}

\item{pointSize}{point size (if points = TRUE)}

\item{StdErr}{maximum standard error to which estimates are displayed}

\item{rangex}{range of longitude values (x axis limits)}

\item{rangey}{range of latitude values (y axis limits)}

\item{rangez}{range of estimated values (z axis limits)}

\item{limitz}{restrict range of z (possible values "0-1", "0-100")}

\item{resolution}{spatial grid resolution of displayed (higher is slower but better quality)}

\item{interior}{show only convex hull TRUE/FALSE}

\item{ncol}{number of colors for estimates}

\item{colors}{color scheme of estimates from RColorBrewer. defaults to "RdYlGn"}

\item{reverseColors}{inverse colour scheme}

\item{colorsP}{color scale of location marks}

\item{centerMap}{center of map, one of "Europe" and "Pacific"}

\item{pointShape}{pch shape of location marks}

\item{showScale}{show colour scale}

\item{textLabels}{text labels}

\item{pointLabels}{point size labels}

\item{pointColLabels}{point colour labels}

\item{fontSize}{font size}

\item{showModel}{show model}

\item{fontType}{font type}

\item{fontCol}{font color}

\item{mask}{boolean: Show output within range of points}

\item{maskRadius}{numeric: Show output within range of points in km}

\item{centerX}{optional. longitude value for fixed spatial
point estimate returns estimate instead of map.}

\item{centerY}{optional. latitude value for fixed spatial
point estimate returns estimate instead of map.}

\item{Radius}{radius of fixed spatial point estimate}

\item{pColor}{color of location marks drawn}

\item{dataCenter}{optional data.frame with two columns (longitude / latitude) for batch estimates
of a series of spatial points}

\item{RadiusBatch}{radius of batch spatial point estimates}

\item{terrestrial}{show only estimates on land masses (1), oceans (-1) or all (0)}

\item{grid}{show coordinate grid TRUE/FALSE}

\item{arrow}{display north arrow TRUE/FALSE}

\item{scale}{display scale TRUE/FALSE}

\item{mapType}{Type of map. "Map" for simple estimates. "speed" for gradient
(mainly for spread model)}

\item{titleMain}{show main title}

\item{titleScale}{show scale title}

\item{setAxisLabels}{show axis/main/scale labels}

\item{mainLabel}{main label}

\item{yLabel}{y lab label}

\item{xLabel}{y lab label}

\item{northSize}{size of north arrow}

\item{scalSize}{size of scale}

\item{scaleX}{scale x orientation}

\item{scaleY}{scale y orientation}

\item{NorthX}{north arrow x orientation}

\item{NorthY}{north arrow y orientation}

\item{scLabel}{scale lab label}

\item{AxisSize}{axis title font size}

\item{AxisLSize}{axis label font size}

\item{cluster}{show clusters}

\item{clusterCol}{Cluster colors}

\item{pointDat}{data frame of points to add to plot}

\item{MinMax}{Min or Max, only for spread model}

\item{nMin}{Number of minima to compare, only for spread model}

\item{minDist}{Distance between minima/maxima, only for spread model}

\item{showMinOnMap}{Show minima on map yes/no, only for spread model if maptype = "Minima/Maxima"}
}
\description{
Plots map of a spatial average or spread model from estimateMap() or estimateMapSpread() functions
}
