% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMap.R
\name{plotMap3D}
\alias{plotMap3D}
\title{Plots time slice map of a spatio-temporal model from estimateMap3D() function}
\usage{
plotMap3D(
  model,
  time,
  IndSelect = NULL,
  estType = "Mean",
  estQuantile = 0.9,
  points = TRUE,
  pointSize = 1,
  StdErr = Inf,
  rangex = range(model$data$Longitude, na.rm = TRUE),
  rangey = range(model$data$Latitude, na.rm = TRUE),
  rangez = NULL,
  limitz = "none",
  centerMap = "Europe",
  addU = 0,
  pointShape = 4,
  pColor = "#000000",
  colorsP = NULL,
  showModel = TRUE,
  textLabels = NULL,
  pointLabels = NULL,
  pointColLabels = NULL,
  fontSize = 11,
  mask = FALSE,
  maskRadius = 100,
  showScale = TRUE,
  fontType = "sans",
  fontCol = "black",
  resolution = 100,
  interior = TRUE,
  ncol = 10,
  colors = "RdYlGn",
  reverseColors = FALSE,
  centerX = NA,
  centerY = NA,
  Radius = NA,
  dataCenter = NULL,
  RadiusBatch = 100,
  terrestrial = 1,
  grid = TRUE,
  arrow = TRUE,
  scale = TRUE,
  titleMain = TRUE,
  titleScale = TRUE,
  setAxisLabels = FALSE,
  mainLabel = "",
  yLabel = "Latitude",
  xLabel = "Longitude",
  scLabel = "",
  northSize = 0.2,
  scalSize = 0.1,
  scaleX = 0,
  scaleY = 0.1,
  NorthX = 0.025,
  NorthY = 0.925,
  AxisSize = 1,
  AxisLSize = 1,
  cluster = FALSE,
  clusterAll = "0",
  clusterCol = "Set1",
  pointDat = NULL,
  plotRetNull = FALSE
)
}
\arguments{
\item{model}{return object of a spatio-temporal model from estimateMap3D() function}

\item{time}{time slice value for map}

\item{IndSelect}{for categorical model: selected category}

\item{estType}{one of "Mean", "SE" and "Quantile". defaults to "Mean"}

\item{estQuantile}{quantile (only applicable if estType = "Quantile")}

\item{points}{show points on map TRUE/FALSE}

\item{pointSize}{point size (if points = TRUE)}

\item{StdErr}{maximum standard error to which estimates are displayed}

\item{rangex}{range of longitude values (x axis limits)}

\item{rangey}{range of latitude values (y axis limits)}

\item{rangez}{range of estimated values (z axis limits)}

\item{limitz}{restrict range of z (possible values "0-1", "0-100")}

\item{centerMap}{center of map, one of "Europe" and "Pacific"}

\item{addU}{numeric: years of added uncertainty for time sliced 2D-convex hull}

\item{pointShape}{pch shape of location marks}

\item{pColor}{color of location marks drawn}

\item{colorsP}{color scale of location marks}

\item{showModel}{show model}

\item{textLabels}{text labels}

\item{pointLabels}{point size labels}

\item{pointColLabels}{point colour labels}

\item{fontSize}{font size}

\item{mask}{boolean: Show output within range of points}

\item{maskRadius}{numeric: Show output within range of points in km}

\item{showScale}{show colour scale}

\item{fontType}{font type}

\item{fontCol}{font color}

\item{resolution}{spatial grid resolution of displayed (higher is slower but better quality)}

\item{interior}{show only convex hull 0 "none", 1 "3D", 2 "time sliced 2D"}

\item{ncol}{number of colors for estimates}

\item{colors}{color scheme of estimates from RColorBrewer. defaults to "RdYlGn"}

\item{reverseColors}{inverse colour scheme}

\item{centerX}{optional. longitude value for fixed spatial
point estimate returns estimate instead of map.}

\item{centerY}{optional. latitude value for fixed spatial
point estimate returns estimate instead of map.}

\item{Radius}{radius of fixed spatial point estimate for batch estimates
of a series of spatial points}

\item{dataCenter}{optional data.frame with three columns (longitude / latitude / time)}

\item{RadiusBatch}{radius of batch spatial point estimates}

\item{terrestrial}{show only estimates on land masses (1), oceans (-1) or all (0)}

\item{grid}{show coordinate grid TRUE/FALSE}

\item{arrow}{display north arrow TRUE/FALSE}

\item{scale}{display scale TRUE/FALSE}

\item{titleMain}{show main title}

\item{titleScale}{show scale title}

\item{setAxisLabels}{show axis/main/scale labels}

\item{mainLabel}{main label}

\item{yLabel}{y lab label}

\item{xLabel}{y lab label}

\item{scLabel}{scale lab label}

\item{northSize}{size of north arrow}

\item{scalSize}{size of scale}

\item{scaleX}{scale x orientation}

\item{scaleY}{scale y orientation}

\item{NorthX}{north arrow x orientation}

\item{NorthY}{north arrow y orientation}

\item{AxisSize}{axis title font size}

\item{AxisLSize}{axis label font size}

\item{cluster}{show clusters}

\item{clusterAll}{show all cluster points}

\item{clusterCol}{Cluster colors}

\item{pointDat}{data frame of points to add to plot}

\item{plotRetNull}{return predictions}
}
\description{
Plots time slice map of a spatio-temporal model from estimateMap3D() function
}
