% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMap.R
\name{plotTimeCourse}
\alias{plotTimeCourse}
\title{Plots time course map of a spatio-temporal model from estimateMap3D() function}
\usage{
plotTimeCourse(
  model,
  independent = "",
  trange = range(model$data$Date),
  resolution = 500,
  centerX = NA,
  centerY = NA,
  Radius = NA,
  rangey = NULL,
  seType = "2",
  pointDat = NULL,
  pointsTime = FALSE,
  returnPred = FALSE,
  intTime = FALSE,
  rangePointsTime = 500,
  limitz = NULL
)
}
\arguments{
\item{model}{return object of a spatio-temporal model from estimateMap3D() function}

\item{independent}{name of independent variable shown in plot}

\item{trange}{range of longitude values (x axis limits)}

\item{resolution}{temporal grid resolution of displayed (higher is slower but better quality)}

\item{centerX}{longitude value to display time course plot for}

\item{centerY}{latitude value to display time course plot for}

\item{Radius}{radius of fixed spatial point estimate}

\item{rangey}{2-element vector of time interval to show}

\item{seType}{setype}

\item{pointDat}{add points/lines to plot}

\item{pointsTime}{should nearby points be plotted?}

\item{returnPred}{should the prediction be returned?}

\item{intTime}{should uncertainty intervals of nearby points be plotted?}

\item{rangePointsTime}{range of nearby points in km}

\item{limitz}{z limit range}
}
\description{
Plots time course map of a spatio-temporal model from estimateMap3D() function
}
