% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{pointColourServer}
\alias{pointColourServer}
\title{server funtion of leaflet point settings module}
\usage{
pointColourServer(id, loadedData, apply)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{loadedData}{(reactive) loaded data}

\item{apply}{(reactive) apply button input}
}
\description{
server funtion of leaflet point settings module
}
