% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{pointColourUI}
\alias{pointColourUI}
\title{ui function of leaflet point settings module}
\usage{
pointColourUI(id)
}
\arguments{
\item{id}{namespace}
}
\description{
ui function of leaflet point settings module
}
