% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{pointSizeServer}
\alias{pointSizeServer}
\title{server funtion of leaflet point settings module}
\usage{
pointSizeServer(id, loadedData)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{loadedData}{(reactive) loaded data}
}
\description{
server funtion of leaflet point settings module
}
