% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{pointSymbolUI}
\alias{pointSymbolUI}
\title{ui function of leaflet point symbol settings module}
\usage{
pointSymbolUI(id)
}
\arguments{
\item{id}{namespace}
}
\description{
ui function of leaflet point symbol settings module
}
