% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictEstimates.R
\name{predict_bayes}
\alias{predict_bayes}
\title{Predict estimates for Bayesian spatio-temporal model}
\usage{
predict_bayes(
  model,
  XPred,
  time = NULL,
  estType = "Mean",
  estQuantile = 0.9,
  sdValue = NULL,
  minVal = NULL,
  maxVal = NULL
)
}
\arguments{
\item{model}{Model object}

\item{XPred}{Prediction grid (data.frame)}

\item{time}{Time slice}

\item{estType}{Estimate type (e.g. "Mean", "1 SE", "Quantile", etc.)}

\item{estQuantile}{Quantile value (if applicable)}

\item{sdValue}{Standard deviation value for uncertainty intervals (if NULL, credible intervals are computed)}

\item{minVal}{Minimum value for predictions (for truncation)}

\item{maxVal}{Maximum value for predictions (for truncation)}
}
\value{
Data frame with predictions and uncertainties
}
\description{
Predict estimates for Bayesian spatio-temporal model
}
