% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictEstimates.R
\name{predict_gam}
\alias{predict_gam}
\title{Predict estimates for KDE spatio-temporal model (GAM)}
\usage{
predict_gam(
  model,
  XPred,
  time = NULL,
  estType = "Mean",
  estQuantile = 0.9,
  sdValue = NULL
)
}
\arguments{
\item{model}{Model object}

\item{XPred}{Prediction grid (data.frame)}

\item{time}{Time slice}

\item{estType}{Estimate type}

\item{estQuantile}{Quantile value}

\item{sdValue}{Standard deviation value for uncertainty intervals (if NULL, credible intervals are computed)}
}
\value{
Data frame with predictions and uncertainties
}
\description{
Predict estimates for KDE spatio-temporal model (GAM)
}
