% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-predictEstimates.R
\name{predict_gamm}
\alias{predict_gamm}
\title{Predict estimates for frequentist spatio-temporal model (GAMM)}
\usage{
predict_gamm(
  model,
  XPred,
  time = NULL,
  estType = "Mean",
  estQuantile = 0.9,
  sdValue = NULL,
  minVal = NULL,
  maxVal = NULL
)
}
\arguments{
\item{model}{Model object}

\item{XPred}{Prediction grid (data.frame)}

\item{time}{Time slice}

\item{estType}{Estimate type}

\item{estQuantile}{Quantile value}

\item{sdValue}{Standard deviation value for uncertainty intervals (if NULL, credible intervals are computed)}

\item{minVal}{Minimum value for predictions (for truncation)}

\item{maxVal}{Maximum value for predictions (for truncation)}
}
\value{
Data frame with predictions and uncertainties
}
\description{
Predict estimates for frequentist spatio-temporal model (GAMM)
}
