% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeData.R
\name{prepareDataServer}
\alias{prepareDataServer}
\title{Prepare Data Server}
\usage{
prepareDataServer(id, selectedData)
}
\arguments{
\item{id}{id of module}

\item{selectedData}{(list) list of data to be merged}
}
\description{
Server function of the module
}
