% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-estimateMap.R
\name{prepareDate}
\alias{prepareDate}
\title{Prepare Date}
\usage{
prepareDate(data, DateOne, DateTwo, DateType, dateUnc, useMaxUnc = TRUE)
}
\arguments{
\item{data}{data.frame: data}

\item{DateOne}{character: name of date variable 1 (lower interval point / mean / single point)}

\item{DateTwo}{character: name of date variable 2 (upper interval point / sd / )}

\item{DateType}{character: one of "Interval", "Mean + 1 SD uncertainty" and "Single Point"}

\item{dateUnc}{character: one of "uniform", "normal", "point"}

\item{useMaxUnc}{(logical) True if max uncertainty should be used.}
}
\description{
Adds new columns 'Date' and 'Uncertainty' that are used in the model dependent on user inputs.
}
