% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-convertLatLong.R
\name{putNewLatLongToBegin}
\alias{putNewLatLongToBegin}
\title{Put Lat Long To Begin}
\usage{
putNewLatLongToBegin(data, Longitude, Latitude)
}
\arguments{
\item{data}{data.frame: data}

\item{Longitude}{character: name of longitude variable}

\item{Latitude}{character: name of latitude variable}
}
\description{
If not already present, than adds new columns "id", latitude" and "longitude" to the beginning of
the dataframe. They are expected in the interactive map
}
