% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-queryData.R
\name{queryDataServer}
\alias{queryDataServer}
\title{Query Data Server}
\usage{
queryDataServer(id, mergeList)
}
\arguments{
\item{id}{id of module}

\item{mergeList}{(list) list of data to be merged}
}
\description{
Server function of the qery data module
}
