% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-processCoordinateData.R
\name{removeDataOutsideRestriction}
\alias{removeDataOutsideRestriction}
\title{Remove Data Outside Restriction}
\usage{
removeDataOutsideRestriction(data, Latitude, Longitude, restriction)
}
\arguments{
\item{data}{(data.frame) data containing columns "Latitude" and "Longitude"}

\item{Latitude}{(character) column name of Latitude column}

\item{Longitude}{(character) column name of Longitude column}

\item{restriction}{(numeric) restriction in the form c(minLatitude, maxLatitude, minLongitude, maxLongitude)}
}
\value{
(data.frame) data containing columns "Latitude" and "Longitude" with data outside of restriction removed
}
\description{
Remove all data outside of the restriction
}
