% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeData.R
\name{removeDuplicatedCols}
\alias{removeDuplicatedCols}
\title{Remove Duplicated Cols}
\usage{
removeDuplicatedCols(df, isTest = FALSE)
}
\arguments{
\item{df}{(data.frame) dataframe with column names}

\item{isTest}{(logical) if TRUE no alert will pop up}
}
\description{
Skip columns whose column names occur twice, keep only that column of the first occurrence of
the column name.
}
