% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{removeOldIcons}
\alias{removeOldIcons}
\title{Remove Old Icons}
\usage{
removeOldIcons(pattern, path = "www")
}
\arguments{
\item{pattern}{pattern of the files}

\item{path}{path to the folder that contains the files}
}
\description{
Remove files of previous icons
}
