% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeData.R
\name{renameColumnsServer}
\alias{renameColumnsServer}
\title{Rename Columns Server}
\usage{
renameColumnsServer(id, columnNames)
}
\arguments{
\item{id}{id of module}

\item{columnNames}{(reactive) column names}
}
\description{
Server function of the module
}
