% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-savedMapsModule.R
\name{savedMapsExportServer}
\alias{savedMapsExportServer}
\title{Saved Maps Export Server}
\usage{
savedMapsExportServer(id, savedMaps)
}
\arguments{
\item{id}{id of module}

\item{savedMaps}{saved maps}
}
\description{
Server function of the module
}
