% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-savedMapsModule.R
\name{savedMapsTab}
\alias{savedMapsTab}
\title{server function of saved maps module}
\usage{
savedMapsTab(input, output, session, savedMaps)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{savedMaps}{saved maps}
}
\description{
server function of saved maps module
}
