% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-savedMapsModule.R
\name{savedMapsTabUI}
\alias{savedMapsTabUI}
\title{ui function of saved maps module}
\usage{
savedMapsTabUI(id, title = "")
}
\arguments{
\item{id}{namespace}

\item{title}{title in tab}
}
\description{
ui function of saved maps module
}
