% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-plotMap.R
\name{selectClusterGrouping}
\alias{selectClusterGrouping}
\title{Select Cluster Grouping}
\usage{
selectClusterGrouping(data, cluster, clusterResults)
}
\arguments{
\item{data}{(data.frame) data if \code{cluster == TRUE} it must contain columns "spatial_cluster"
and "temporal_group"}

\item{cluster}{(logical) if TRUE, a column "cluster" is added to data}

\item{clusterResults}{(integer) either \code{0} or \code{1}, using "temporal_group" or
"spatial_cluster" for the column "cluster", respectively.}
}
\value{
(data.frame) data containing the column "cluster" if \code{cluster == TRUE}, else
 just the input \code{data}
}
\description{
Select Cluster Grouping
}
