% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-interactiveMap_NorthArrow-class.R
\name{set.NorthArrow}
\alias{set.NorthArrow}
\title{Set North Arrow on a Leaflet Map}
\usage{
\method{set}{NorthArrow}(obj, map, ...)
}
\arguments{
\item{obj}{A NorthArrow object.}

\item{map}{A Leaflet map object.}

\item{...}{Additional arguments (not used).}
}
\value{
A Leaflet map object with the North Arrow added.
}
\description{
Sets the North Arrow on a Leaflet map based on the properties of the NorthArrow object.
}
