% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{set}
\alias{set}
\title{Generic function for setting values in an object}
\usage{
set(obj, ...)
}
\arguments{
\item{obj}{An object of class ScaleBar or similar.}

\item{...}{Additional arguments.}
}
\description{
Generic function for setting values in an object
}
