% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-interactiveMap_ScaleBar-class.R
\name{set.ScaleBar}
\alias{set.ScaleBar}
\title{Set Scale Bar on Leaflet Map}
\usage{
\method{set}{ScaleBar}(obj, map, ...)
}
\arguments{
\item{obj}{An object of class `ScaleBar`.}

\item{map}{A Leaflet map object.}

\item{...}{Additional arguments (not used).}
}
\value{
A Leaflet map object with the scale bar added.
}
\description{
This function sets a scale bar on a Leaflet map based on the properties of a `ScaleBar` object.
}
