% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{setSizeLegend}
\alias{setSizeLegend}
\title{Set Size Legend}
\usage{
setSizeLegend(map, showLegend, title, factor, radius, data)
}
\arguments{
\item{map}{leaflet map}

\item{showLegend}{logical show/hide legend}

\item{title}{legend title}

\item{factor}{factor for size}

\item{radius}{displayed radius}

\item{values}{underlying values that are mapped to size}
}
\description{
Set Size Legend
}
