% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-processCoordinateData.R
\name{shiftDataToDefaultRestriction}
\alias{shiftDataToDefaultRestriction}
\title{Shift Data To Default Restriction}
\usage{
shiftDataToDefaultRestriction(data)
}
\arguments{
\item{data}{(data.frame) data containing columns "Latitude" and "Longitude"}
}
\value{
(data.frame) data containing columns "Latitude" and "Longitude" shifted to the default restriction
}
\description{
Shift data such that it is in the range of -180 to 180 and -90 to 90
}
