% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-interactiveMap.R
\name{showIDPopup}
\alias{showIDPopup}
\title{Show a popup at the given location}
\usage{
showIDPopup(dat, id, lat, lng)
}
\arguments{
\item{dat}{dat contains data to show}

\item{id}{id of what to show}

\item{lat}{lat for popup}

\item{lng}{lng for popup}
}
\description{
Show a popup at the given location
}
