% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-skin.R
\name{setSkin}
\alias{setSkin}
\alias{getSkin}
\alias{allowedSkins}
\title{Skin helper functions}
\usage{
setSkin(skin)

getSkin()

allowedSkins()
}
\arguments{
\item{skin}{one of isomemo / pandora}
}
\description{
Skin helper functions
}
