% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelResults3D.R
\name{sliderAndNumericInputServer}
\alias{sliderAndNumericInputServer}
\title{Slider And Input Server}
\usage{
sliderAndNumericInputServer(id)
}
\arguments{
\item{id}{id of module}
}
\description{
Server function of the Slider And Input module
}
