% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-ELEMENTS.R
\name{sliderAndNumericInputUI}
\alias{sliderAndNumericInputUI}
\title{Slider And Input UI}
\usage{
sliderAndNumericInputUI(id, label, min, max, value, step)
}
\arguments{
\item{id}{id of module}

\item{label}{label}

\item{min}{(numeric) minumum}

\item{max}{(numeric) maximum}

\item{value}{(numeric) default value}

\item{step}{(numeric) step}
}
\description{
UI of the Slider And Input module
}
