% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelMap_ELEMENTS.R
\name{smoothingServer}
\alias{smoothingServer}
\title{Smoothing Server}
\usage{
smoothingServer(
  input,
  output,
  session,
  ns,
  title = "Number of Basis Functions (Smoothing)",
  map3D = FALSE,
  label = "No. of time basis functions"
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{ns}{namespace}

\item{title}{title of the modal}

\item{map3D}{logical, if TRUE, the map is 3D (contains the time dimension)}

\item{label}{if `map3D == TRUE` label of the slider for the number of time basis functions}
}
\description{
Smoothing Server
}
