% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-ELEMENTS.R
\name{smoothingUI}
\alias{smoothingUI}
\title{Smoothing UI}
\usage{
smoothingUI(
  ns,
  label_slider = "No. of basis functions",
  label_info = "Basis functions",
  min = 20,
  max = 1000,
  value = 70,
  step = 10
)
}
\arguments{
\item{ns}{namespace}

\item{label_slider}{label of the slider}

\item{label_info}{label of the info button}

\item{min}{minimum value of the slider}

\item{max}{maximum value of the slider}

\item{value}{default value of the slider}

\item{step}{step of the slider}
}
\description{
Smoothing UI
}
