% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeData.R
\name{splitColumnsServer}
\alias{splitColumnsServer}
\title{Split Columns Server}
\usage{
splitColumnsServer(id, preparedData)
}
\arguments{
\item{id}{id of module}

\item{preparedData}{(reactive) (reactive) selected data, possibly already modified}
}
\description{
Server function of the module
}
