% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{timeAndMapSectionServer}
\alias{timeAndMapSectionServer}
\title{Time And Map Section Server}
\usage{
timeAndMapSectionServer(id, dateExtent)
}
\arguments{
\item{id}{id of module}

\item{dateExtent}{(reactiveValues) date range parameters}
}
\description{
Server function of the module
}
