% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{timeAndMapSectionUI}
\alias{timeAndMapSectionUI}
\title{Time And Map Section UI}
\usage{
timeAndMapSectionUI(id, label)
}
\arguments{
\item{id}{id of module}

\item{label}{label}
}
\description{
UI of the module
}
