% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{updateDataOnLeafletMap}
\alias{updateDataOnLeafletMap}
\title{Update Data On Map}
\usage{
updateDataOnLeafletMap(map, isoData, leafletPointValues)
}
\arguments{
\item{map}{reactive leaflet map object}

\item{isoData}{reactive isoData data}

\item{leafletPointValues}{reactive settings for points on map}
}
\description{
Update Data On Map
}
