% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotExport.R
\name{writeGraphics}
\alias{writeGraphics}
\title{Write Graphics}
\usage{
writeGraphics(exportType, plot, filename, width, height)
}
\arguments{
\item{exportType}{(character) file type of exported plot}

\item{plot}{(object) plot object}

\item{filename}{the output file path.
    The page number is substituted if a C integer format is included in
    the character string, as in the default.  (Depending on the platform,
    the result should be less
    than \code{PATH_MAX} characters long, and may be truncated if not.
    See \code{\link[grDevices]{pdf}} for further details.)  Tilde expansion
    is performed where supported by the platform.  An input with
    a marked encoding is converted to the native encoding on an error is
    given.}

\item{width}{the width of the device.}

\item{height}{the height of the device.}
}
\description{
Write Graphics
}
