% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-processCoordinateData.R
\name{augmentData}
\alias{augmentData}
\title{Augment Data}
\usage{
augmentData(data, restriction = c(-90, 90, -320, 320))
}
\arguments{
\item{data}{(data.frame) data containing columns "Latitude" and "Longitude"}

\item{restriction}{(numeric) restriction in the form c(minLatitude, maxLatitude, minLongitude, maxLongitude)}
}
\value{
(data.frame) augmented data
}
\description{
Data augmentation to add data at the borders of the map
}
