% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-processCoordinateData.R
\name{centerData}
\alias{centerData}
\title{Center Data}
\usage{
centerData(data, center = c("Europe", "Pacific"))
}
\arguments{
\item{data}{(data.frame) data containing column "Longitude"}

\item{center}{(character) center to shift data to, either "Europe" or "Pacific"}
}
\value{
(data.frame) data shifted to the center
}
\description{
Transfer data to the center of the map, either Europe or Pacific by adding or substracting 360°
from the Longitude. If "Europe" longitudes are transposed to the range (-180, 180),  if
"Pacific" longitudes are transposed to the range (0, 360).
}
