% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{centerEstimateServer}
\alias{centerEstimateServer}
\title{Center Estimate Server}
\usage{
centerEstimateServer(id, predictions, mapType = reactiveVal("Map"))
}
\arguments{
\item{id}{namespace id}

\item{predictions}{(reactive) all predictions}

\item{mapType}{(reactive) type of plot, either "Map", "Time course", "Time intervals by temporal group or cluster",
"Spread", "Speed", "Minima/Maxima"}
}
\description{
Backend for center estimate module
}
