% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-interactiveMap.R
\name{draw}
\alias{draw}
\title{draw Interactive Map}
\usage{
draw(
  isoData,
  zoom = 5,
  type = "1",
  northArrowPosition = "none",
  scalePosition = "none",
  center = NULL,
  bounds = NULL
)
}
\arguments{
\item{isoData}{isoData data}

\item{zoom}{zoom}

\item{type}{map type}

\item{northArrowPosition}{position of north arrow}

\item{scalePosition}{position of scale}

\item{center}{where to center map (list of lat and lng)}

\item{bounds}{map bounds (list of north, south, east, west)}
}
\description{
draw Interactive Map
}
