% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-leafletMapHelpers.R
\name{draw}
\alias{draw}
\title{draw Interactive Map}
\usage{
draw(
  isoData,
  zoom = 5,
  type = "1",
  northArrow = FALSE,
  northArrowPosition = "bottomright",
  scale = FALSE,
  scalePosition = "topleft",
  logoPosition = NA,
  center = NULL
)
}
\arguments{
\item{isoData}{isoData data}

\item{zoom}{zoom}

\item{type}{map type}

\item{northArrow}{show north arrow?}

\item{northArrowPosition}{position of north arrow}

\item{scale}{show scale?}

\item{scalePosition}{position of scale}

\item{logoPosition}{character position of logo if selected, else NA}

\item{center}{where to center map (list of lat and lng)}
}
\description{
draw Interactive Map
}
