% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-interactiveMap.R
\name{drawIcons}
\alias{drawIcons}
\title{Draw Icons on Interactive Map}
\usage{
drawIcons(
  map,
  northArrowPosition = "none",
  northArrowSize = 80,
  scalePosition = "none",
  scaleSize = 100
)
}
\arguments{
\item{map}{leaflet map}

\item{northArrowPosition}{position of north arrow}

\item{scalePosition}{position of scale}

\item{northArrow}{show north arrow?}

\item{scale}{show scale?}
}
\description{
Draw Icons on Interactive Map
}
