% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-savedMapsModule.R
\name{filterCoordSquare}
\alias{filterCoordSquare}
\title{Filter Coord Square}
\usage{
filterCoordSquare(fullGrid, lat, long, radius)
}
\arguments{
\item{fullGrid}{grid of coordinates}

\item{lat}{(numeric) latitude}

\item{long}{(numeric) longitude}

\item{radius}{(numeric) radius}
}
\description{
Filter full grid of coordinates to square area
}
