% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-findDuplicates.R
\name{findDuplicates}
\alias{findDuplicates}
\title{Finds duplicates in data using user specified similarity rules}
\usage{
findDuplicates(data, userSimilaritySelection, addColumn, keepFirst)
}
\arguments{
\item{data}{dataframe in which duplicates are searched for}

\item{userSimilaritySelection}{dataframe containing similarity rules for each column to be considered}

\item{addColumn}{logical should a column with duplicate row indices be added}

\item{keepFirst}{logical keep first (TRUE) or last (FALSE) duplicate row}
}
\description{
Finds duplicates in data using user specified similarity rules
}
