% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{formatTimeCourseServer}
\alias{formatTimeCourseServer}
\title{Format Time Course Server}
\usage{
formatTimeCourseServer(id)
}
\arguments{
\item{id}{namespace id}
}
\description{
Backend for formatting of the time course plot
}
