% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-mapOperations.R
\name{getBinaryPredictions}
\alias{getBinaryPredictions}
\title{Get Binary Predictions}
\usage{
getBinaryPredictions(XPred1, XPred2)
}
\arguments{
\item{XPred1}{(data.frame) prediction matrix of first map}

\item{XPred2}{(data.frame) prediction matrix of second map}
}
\value{
(data.frame) prediction matrix of resulting map
}
\description{
Get Binary Predictions
}
