% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{getZvalues}
\alias{getZvalues}
\title{Get Z Values}
\usage{
getZvalues(estimationType, model, mapType)
}
\arguments{
\item{estimationType}{(character) type of estimate}

\item{model}{(list) model output}

\item{mapType}{(character) type of map, either "Map" or "Time course"}
}
\description{
Get Z Values
}
