% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{getZvalues}
\alias{getZvalues}
\title{Get Z Values}
\usage{
getZvalues(estimationType, model)
}
\arguments{
\item{estimationType}{(character) type of estimate}

\item{model}{(list) model output}
}
\description{
Get Z Values
}
