% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{getZValuesKernel}
\alias{getZValuesKernel}
\title{Get Z Values Kernel}
\usage{
getZValuesKernel(
  estimationType,
  model,
  mapType,
  factor = 1.25,
  IndSelect = NULL
)
}
\arguments{
\item{estimationType}{(character) type of estimate}

\item{model}{(list) model output}

\item{mapType}{(character) type of map, either "Map" or "Time course"}

\item{factor}{(numeric) factor applied to estimates}

\item{IndSelect}{(character) select category in case of categorical model}
}
\description{
Get Z Values Kernel
}
