% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-importData.R
\name{importDataUI}
\alias{importDataUI}
\alias{importData}
\title{Data import module}
\usage{
importDataUI(id, label = "Import Data")

importData(
  input,
  output,
  session,
  rowNames = NULL,
  colNames = NULL,
  customWarningChecks = list(),
  customErrorChecks = list()
)
}
\arguments{
\item{id}{id of module}

\item{label}{label of button}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{rowNames}{(reactive) use this for rownames of imported data}

\item{colNames}{(reactive) use this for colnames of imported data}

\item{customWarningChecks}{list of reactive functions which will be executed after importing of data.
functions need to return TRUE if check is successful or a character with a warning otherwise.}

\item{customErrorChecks}{list of reactive functions which will be executed after importing of data.
functions need to return TRUE if check is successful or a character with a warning otherwise.}
}
\description{
Displays a button which opens a import dialog when clicked

Backend for data import module
}
