% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletExport.R
\name{leafletExport}
\alias{leafletExport}
\title{server funtion of leaflet map export module}
\usage{
leafletExport(input, output, session, leafletMap, width, height)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{leafletMap}{reactive leaflet map object}

\item{width}{reactive width of map in px}

\item{height}{reactive height of map in px}
}
\description{
server funtion of leaflet map export module
}
