% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletSettings.R
\name{leafletSettings}
\alias{leafletSettings}
\title{server funtion of leaflet settings module}
\usage{
leafletSettings(input, output, session, map)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\description{
server funtion of leaflet settings module
}
