% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-mapDiff.R
\name{mapDiff}
\alias{mapDiff}
\title{server function of model Results module}
\usage{
mapDiff(input, output, session, savedMaps, fruitsData)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{savedMaps}{saved Maps}

\item{fruitsData}{data for export to FRUITS}
}
\description{
server function of model Results module
}
