% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{mapSectionServer}
\alias{mapSectionServer}
\title{Map Section Server}
\usage{
mapSectionServer(id)
}
\arguments{
\item{id}{id of module}
}
\description{
Server function of the module
}
