% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-mapSim.R
\name{mapSim}
\alias{mapSim}
\title{server function of model Results module}
\usage{
mapSim(input, output, session, savedMaps, fruitsData, config)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{savedMaps}{saved Maps}

\item{fruitsData}{data for export to FRUITS}

\item{config}{(list) list of configuration parameters}
}
\description{
server function of model Results module
}
