% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeViaUI.R
\name{mergeViaUIServer}
\alias{mergeViaUIServer}
\title{Merge Via UI Server}
\usage{
mergeViaUIServer(id, mergeList)
}
\arguments{
\item{id}{id of module}

\item{mergeList}{(list) list of data to be merged}
}
\description{
Server function of the merge via UI module
}
