% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-modelResults2DKernel.R
\name{modelResults2DKernel}
\alias{modelResults2DKernel}
\title{server function of model Results module}
\usage{
modelResults2DKernel(input, output, session, isoData, savedMaps, fruitsData)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{isoData}{data}

\item{savedMaps}{saved Maps}

\item{fruitsData}{data for export to FRUITS}
}
\description{
server function of model Results module
}
