% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-mergeData.R
\name{prepareDataServer}
\alias{prepareDataServer}
\title{Prepare Data Server}
\usage{
prepareDataServer(id, selectedData, nameOfSelected)
}
\arguments{
\item{id}{id of module}

\item{selectedData}{(reactive) selected data}

\item{nameOfSelected}{(reactive) filename of selected data}
}
\description{
Server function of the module
}
