% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-interactiveMap.R
\name{setColorLegend}
\alias{setColorLegend}
\title{Add Colour Legend}
\usage{
setColorLegend(map, showLegend, values)
}
\arguments{
\item{map}{leaflet map}

\item{showLegend}{logical show/hide legend}

\item{values}{possible values that can be mapped, e.g. isoData$source}
}
\description{
Add Colour Legend
}
