% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{setColorLegend}
\alias{setColorLegend}
\title{Set Colour Legend}
\usage{
setColorLegend(map, showLegend, title, pal, values)
}
\arguments{
\item{map}{leaflet map}

\item{showLegend}{logical show/hide legend}

\item{title}{legend title}

\item{pal}{colour palette}

\item{values}{possible values that can be mapped, e.g. isoData$source}
}
\description{
Set Colour Legend
}
