% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-leafletPointSettings.R
\name{setSizeLegend}
\alias{setSizeLegend}
\title{Set Size Legend}
\usage{
setSizeLegend(map, title, values)
}
\arguments{
\item{map}{leaflet map}

\item{title}{legend title}

\item{values}{possible values that can be mapped, e.g. isoData$source}
}
\description{
Set Size Legend
}
