% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{sliderAndNumericInputServer}
\alias{sliderAndNumericInputServer}
\title{Slider And Input Server}
\usage{
sliderAndNumericInputServer(id, value, min, max, step)
}
\arguments{
\item{id}{id of module}

\item{value}{value of input}

\item{min}{min of input}

\item{max}{max of input}

\item{step}{step of input}
}
\description{
Server function of the Slider And Input module
}
