% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-ELEMENTS.R
\name{smoothingServer}
\alias{smoothingServer}
\title{Smoothing Server}
\usage{
smoothingServer(
  input,
  output,
  session,
  ns,
  title = "Number of Basis Functions (Smoothing)"
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{ns}{namespace}

\item{title}{title of the modal}
}
\description{
Smoothing Server
}
