% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modelMap_UIHelpers.R
\name{zScaleServer}
\alias{zScaleServer}
\title{Z Scale Server}
\usage{
zScaleServer(
  id,
  Model,
  fixCol,
  estimationTypeChoices,
  restrictOption,
  zValuesFun,
  zValuesFactor,
  mapType = reactive("Map"),
  IndSelect = NULL
)
}
\arguments{
\item{id}{id of module}

\item{Model}{(reactive) model output}

\item{fixCol}{(reactive) user input if columns should be fixed, TRUE or FALSE}

\item{estimationTypeChoices}{(reactive) named characters of choices of estimation types}

\item{restrictOption}{(reactive) either "hide" or "show". If "show" than add user input to
restrict the z scale.}

\item{zValuesFun}{(reactive) function to extract zValues, either getZValues or getZValuesKernel}

\item{zValuesFactor}{(numeric) factor applied to zValues}

\item{mapType}{(character)}

\item{IndSelect}{(character) select category in case of categorical model}
}
\description{
Server function of the module
}
