% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{zScaleServer}
\alias{zScaleServer}
\title{Z Scale Server}
\usage{
zScaleServer(id, restrictOption, zValues)
}
\arguments{
\item{id}{id of module}

\item{restrictOption}{(reactive) either "hide" or "show". If "show" than add user input to
restrict the z scale.}

\item{zValues}{(reactive) list with default values for valueMin, valueMax, min, max}
}
\description{
Server function of the module
}
