% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-modules-modelResultHelpers.R
\name{zScaleServer}
\alias{zScaleServer}
\title{Z Scale Server}
\usage{
zScaleServer(id, Model, fixCol)
}
\arguments{
\item{id}{id of module}
}
\description{
Server function of the module
}
